Restructure files paths in repository so it's easier to see what's what.
