+++ legacy/noteslib.js
