+++ src/service-worker/static-resources.js
