+++ src/webpack.config.js
@@ -2,11 +2,11 @@ import { URL } from 'url';
 export default {
   entry: {
     client: './client/index.js',
-    serviceworker: './client/service-worker/index.js',
+    serviceworker: './service-worker/index.js',
   },
   output: {
     filename: '[name].js',
-    path: new URL('./static/', import.meta.url).pathname,
+    path: new URL('./resources/', import.meta.url).pathname,
   },
   mode: 'production',
 };
